+++ examples/inference_cookbook/__init__.py
@@ -1,0 +1,1 @@
+"""Inference cookbook example package."""